Refactor the admin side to use the new vuestore methods
Refactored the admin side to use the new vuestore getters, actions
and mutations.

Fixed a few UI issues including making the coin sign consistent acros the
whole platform.

Kept only a single navbar and made it a bit narrower as expected.

Add a component which will be used to remove html attributes in
case the string of an element contains html tags